Updated filters to be dynamic. fix #52. Moved plugins t own subfolders
