+++ puzzle/blueprints/variants/templates/sv_variants.html
@@ -13,7 +13,7 @@
 
 			<div id="view-filters" class="panel-collapse collapse">
 				<div class="panel-body">
-					{{ filters_form(case_id, inheritance_models, consequences, filters) }}
+					{{ filters_form(case_id, app, inheritance_models, consequences, filters, sv_types) }}
 				</div>
 			</div>
 		</div>
