+++ puzzle/plugins/gemini/__init__.py
@@ -1,0 +1,1 @@
+from .gemini_plugin import GeminiPlugin
