+++ puzzle/plugins/gemini/gemini_plugin.py
