+++ puzzle/plugins/vcf/__init__.py
@@ -1,0 +1,3 @@
+from .variant_mixin import VariantMixin
+from .case_mixin import CaseMixin
+from .vcf import VcfPlugin
