+++ puzzle/plugins/vcf/vcf.py
@@ -1,0 +1,57 @@
+# -*- coding: utf-8 -*-
+import logging
+
+
+from puzzle.plugins import Plugin
+from . import VariantMixin, CaseMixin
+
+logger = logging.getLogger(__name__)
+
+
+class VcfPlugin(VariantMixin, CaseMixin, Plugin):
+    """docstring for Plugin"""
+
+    def __init__(self):
+        super(VcfPlugin, self).__init__()
+        self.db = None
+        self.individuals = None
+        self.case_obj = None
+        
+
+    def init_app(self, app):
+        """Initialize plugin via Flask."""
+        logger.debug("Updating root path to {0}".format(
+            app.config['PUZZLE_ROOT']
+        ))
+        self.root_path = app.config['PUZZLE_ROOT']
+        logger.debug("Updating pattern to {0}".format(
+            app.config['PUZZLE_PATTERN']
+        ))
+        self.pattern = app.config['PUZZLE_PATTERN']
+        
+        self.mode = app.config['PUZZLE_MODE']
+        logger.info("Setting ")
+        logger.debug("Setting can_filter_gene to 'True'")
+        self.can_filter_gene = True
+        if self.mode == 'sv':
+            logger.debug("Setting can_filter_sv to 'True'")
+            self.can_filter_sv = True
+            logger.debug("Setting can_filter_sv_len to 'True'")
+            self.can_filter_sv_len = True
+        else:
+            logger.debug("Setting can_filter_frequency to 'True'")
+            self.can_filter_frequency = True
+            logger.debug("Setting can_filter_cadd to 'True'")
+            self.can_filter_cadd = False
+            logger.debug("Setting can_filter_consequence to 'True'")
+            self.can_filter_consequence = False
+            logger.debug("Setting can_filter_inheritance to 'True'")
+            self.can_filter_inheritance = False
+
+        if app.config.get('FAMILY_FILE'):
+            #If ped file we know there is only one vcf
+            self.db = app.config['PUZZLE_ROOT'].replace('/', '|')
+            self.individuals = self._get_family_individuals(
+                app.config['FAMILY_FILE'], app.config['FAMILY_TYPE'])
+            self.case_obj = self._get_family_case(self.individuals)
+
